break up dependency chain
